+++ Sources/SolarDesignSystem/Components/VisualEffectView/VisualEffect.swift
@@ -26,29 +26,35 @@
 
 #if !os(watchOS) && canImport(UIKit)
 import UIKit
+#elseif os(macOS)
+import AppKit
 #endif
 
+@available(macOS 11, iOS 14.0, tvOS 14.0, *)
+@available(watchOS, unavailable)
 public enum VisualEffect: Hashable {
     
     // MARK: - Cases
     
     /// A visual effect that adapts to the current environment.
-    case adaptive(BlurEffect, VibrancyEffect? = nil)
+    @available(macOS 11, iOS 14.0, tvOS 14.0, *)
+    case adaptive(blurEffect: BlurEffect, vibrancyEffect: VibrancyEffect? = nil, blendingMode: BlendingMode = .withinWindow)
     
     /// A visual effect that uses light colors.
-    case light(BlurEffect, VibrancyEffect? = nil)
+    @available(macOS 11, iOS 14.0, tvOS 14.0, *)
+    case light(blurEffect: BlurEffect, vibrancyEffect: VibrancyEffect? = nil, blendingMode: BlendingMode = .withinWindow)
     
     /// A visual effect that uses dark colors.
-    case dark(BlurEffect, VibrancyEffect? = nil)
+    @available(macOS 11, iOS 14.0, tvOS 14.0, *)
+    case dark(blurEffect: BlurEffect, vibrancyEffect: VibrancyEffect? = nil, blendingMode: BlendingMode = .withinWindow)
     
     // MARK: - Platform Conversion
     
-    #if !os(watchOS) && canImport(UIKit)
+    #if os(iOS)
     
     internal var platformBlurEffect: UIBlurEffect {
-        #if os(iOS)
         switch self {
-        case let .adaptive(blurEffect, _):
+        case let .adaptive(blurEffect, _, _):
             switch blurEffect {
             case .ultraThinMaterial:
                 return UIBlurEffect(style: .systemUltraThinMaterial)
@@ -61,7 +67,7 @@ public enum VisualEffect: Hashable {
             case .chromeMaterial:
                 return UIBlurEffect(style: .systemChromeMaterial)
             }
-        case let .light(blurEffect, _):
+        case let .light(blurEffect, _, _):
             switch blurEffect {
             case .ultraThinMaterial:
                 return UIBlurEffect(style: .systemUltraThinMaterialLight)
@@ -74,7 +80,7 @@ public enum VisualEffect: Hashable {
             case .chromeMaterial:
                 return UIBlurEffect(style: .systemChromeMaterialLight)
             }
-        case let .dark(blurEffect, _):
+        case let .dark(blurEffect, _, _):
             switch blurEffect {
             case .ultraThinMaterial:
                 return UIBlurEffect(style: .systemUltraThinMaterialDark)
@@ -88,43 +94,15 @@ public enum VisualEffect: Hashable {
                 return UIBlurEffect(style: .systemChromeMaterialDark)
             }
         }
-        #elseif os(tvOS)
-        switch self {
-        case let .adaptive(blurEffect, _):
-            switch blurEffect {
-            case .regular:
-                return UIBlurEffect(style: .regular)
-            case .prominent:
-                return UIBlurEffect(style: .prominent)
-            }
-        case let .light(blurEffect, _):
-            switch blurEffect {
-            case .regular:
-                return UIBlurEffect(style: .light)
-            case .prominent:
-                return UIBlurEffect(style: .extraLight)
-            }
-        case let .dark(blurEffect, _):
-            switch blurEffect {
-            case .regular:
-                return UIBlurEffect(style: .dark)
-            case .prominent:
-                return UIBlurEffect(style: .extraDark)
-            }
-        }
-        #endif
     }
     
     internal var platformVibrancyEffect: UIVibrancyEffect? {
         switch self {
-        case let .adaptive(_, vibrancyEffect), let .light(_, vibrancyEffect), let .dark(_, vibrancyEffect):
+        case let .adaptive(_, vibrancyEffect, _), let .light(_, vibrancyEffect, _), let .dark(_, vibrancyEffect, _):
             guard let vibrancyEffect = vibrancyEffect else {
                 return nil
             }
             
-            #if os(tvOS)
-            return UIVibrancyEffect(blurEffect: platformBlurEffect)
-            #else
             switch vibrancyEffect {
             case .default:
                 return UIVibrancyEffect(blurEffect: platformBlurEffect)
@@ -145,10 +123,161 @@ public enum VisualEffect: Hashable {
             case .separator:
                 return UIVibrancyEffect(blurEffect: platformBlurEffect, style: .separator)
             }
-            #endif
         }
     }
     
-    #endif
+    #elseif os(tvOS)
     
+    internal var platformBlurEffect: UIBlurEffect {
+        switch self {
+        case let .adaptive(blurEffect, _, _):
+            switch blurEffect {
+            case .regular:
+                return UIBlurEffect(style: .regular)
+            case .prominent:
+                return UIBlurEffect(style: .prominent)
+            }
+        case let .light(blurEffect, _, _):
+            switch blurEffect {
+            case .regular:
+                return UIBlurEffect(style: .light)
+            case .prominent:
+                return UIBlurEffect(style: .extraLight)
+            }
+        case let .dark(blurEffect, _, _):
+            switch blurEffect {
+            case .regular:
+                return UIBlurEffect(style: .dark)
+            case .prominent:
+                return UIBlurEffect(style: .extraDark)
+            }
+        }
+    }
+    
+    internal var platformVibrancyEffect: UIVibrancyEffect? {
+        switch self {
+        case let .adaptive(_, vibrancyEffect, _), let .light(_, vibrancyEffect, _), let .dark(_, vibrancyEffect, _):
+            guard let _ = vibrancyEffect else {
+                return nil
+            }
+            
+            return UIVibrancyEffect(blurEffect: platformBlurEffect)
+        }
+    }
+    
+    #elseif os(macOS)
+    
+    internal var platformMaterial: NSVisualEffectView.Material {
+        switch self {
+        case let .adaptive(blurEffect, _, _), let .light(blurEffect, _, _), let .dark(blurEffect, _, _):
+            switch blurEffect {
+            case .titlebar:
+                return .titlebar
+            case .selection:
+                return .selection
+            case .menu:
+                return .menu
+            case .popover:
+                return .popover
+            case .sidebar:
+                return .sidebar
+            case .headerView:
+                return .headerView
+            case .sheet:
+                return .sheet
+            case .windowBackground:
+                return .windowBackground
+            case .hudWindow:
+                return .hudWindow
+            case .fullScreenUI:
+                return .fullScreenUI
+            case .toolTip:
+                return .toolTip
+            case .contentBackground:
+                return .contentBackground
+            case .underWindowBackground:
+                return .underWindowBackground
+            case .underPageBackground:
+                return .underPageBackground
+            }
+        }
+    }
+    
+    internal var platformAppearance: NSAppearance {
+        switch self {
+        case let .adaptive(_, vibrancy, _):
+            if let _ = vibrancy {
+                if isHighContrast {
+                    return NSAppearance(named: isDark ? .accessibilityHighContrastVibrantDark: .accessibilityHighContrastVibrantLight)!
+                }
+                return NSAppearance(named: isDark ? .vibrantDark : .vibrantLight)!
+            } else {
+                if isHighContrast {
+                    return NSAppearance(named: isDark ? .accessibilityHighContrastDarkAqua : .accessibilityHighContrastAqua)!
+                }
+                return NSAppearance(named: isDark ? .darkAqua : .aqua)!
+            }
+        case let .light(_, vibrancy, _):
+            if let _ = vibrancy {
+                if isHighContrast {
+                    return NSAppearance(named: .accessibilityHighContrastVibrantLight)!
+                }
+                return NSAppearance(named: .vibrantLight)!
+            } else {
+                if isHighContrast {
+                    return NSAppearance(named: .accessibilityHighContrastAqua)!
+                }
+                return NSAppearance(named: .aqua)!
+            }
+        case let .dark(_, vibrancy, _):
+            if let _ = vibrancy {
+                if isHighContrast {
+                    return NSAppearance(named: .accessibilityHighContrastVibrantDark)!
+                }
+                return NSAppearance(named: .vibrantDark)!
+            } else {
+                if isHighContrast {
+                    return NSAppearance(named: .accessibilityHighContrastDarkAqua)!
+                }
+                return NSAppearance(named: .darkAqua)!
+            }
+        }
+    }
+    
+    internal var platformBlendingMode: NSVisualEffectView.BlendingMode {
+        switch self {
+        case let .adaptive(_, _, blendingMode), let .light(_, _, blendingMode), let .dark(_, _, blendingMode):
+            switch blendingMode {
+            case .withinWindow:
+                return .withinWindow
+            case .behindWindow:
+                return .behindWindow
+            }
+        }
+    }
+    
+    private var isHighContrast: Bool {
+        let highContrastAppearances: [NSAppearance.Name] = [
+            .accessibilityHighContrastAqua,
+            .accessibilityHighContrastDarkAqua,
+            .accessibilityHighContrastVibrantLight,
+            .accessibilityHighContrastVibrantDark
+        ]
+        
+        return highContrastAppearances.contains(NSAppearance.current.name)
+    }
+    
+    private var isDark: Bool {
+        let darkAppearances: [NSAppearance.Name] = [
+            .darkAqua,
+            .accessibilityHighContrastDarkAqua,
+            .vibrantDark,
+            .accessibilityHighContrastVibrantDark
+        ]
+        
+        return darkAppearances.contains(NSAppearance.current.name)
+    }
+    
+    #endif
 }
+