+++ Sources/SolarDesignSystem/Theme/Colors/SystemColor.swift
@@ -42,33 +42,33 @@ public struct SystemColor {
     // MARK: - Labels
     
     /// The color for text labels that contain primary content.
-    @available(iOS 14.0, watchOS 6.0, tvOS 14.0, macOS 10.16, *)
+    @available(iOS 14.0, watchOS 6.0, tvOS 14.0, macOS 11.0, *)
     public var label: Color {
         return colorStorage["label"]!
     }
     
     /// The color for text labels that contain secondary content.
-    @available(iOS 14.0, watchOS 6.0, tvOS 14.0, macOS 10.16, *)
+    @available(iOS 14.0, watchOS 6.0, tvOS 14.0, macOS 11.0, *)
     public var secondaryLabel: Color {
         return colorStorage["secondaryLabel"]!
     }
     
     /// The color for text labels that contain tertiary content.
-    @available(iOS 14.0, tvOS 14.0, macOS 10.16, *)
+    @available(iOS 14.0, tvOS 14.0, macOS 11.0, *)
     @available(watchOS, unavailable)
     public var tertiaryLabel: Color {
         return colorStorage["tertiaryLabel"]!
     }
     
     /// The color for text labels that contain quaternary content.
-    @available(iOS 14.0, tvOS 14.0, macOS 10.16, *)
+    @available(iOS 14.0, tvOS 14.0, macOS 11.0, *)
     @available(watchOS, unavailable)
     public var quaternaryLabel: Color {
         return colorStorage["quaternaryLabel"]!
     }
     
     /// The color for labels that contain primary content layered on top of a tint color.
-    @available(iOS 14.0, tvOS 14.0, watchOS 6.0, macOS 10.16, *)
+    @available(iOS 14.0, tvOS 14.0, watchOS 6.0, macOS 11.0, *)
     public var inverseLabel: Color {
         return colorStorage["inverseLabel"]!
     }
@@ -114,7 +114,7 @@ public struct SystemColor {
     // MARK: - Text
     
     /// The color to use for text.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -123,14 +123,14 @@ public struct SystemColor {
     }
     
     /// The color for placeholder text in controls or text views.
-    @available(iOS 14.0, tvOS 14.0, macOS 10.16, *)
+    @available(iOS 14.0, tvOS 14.0, macOS 11.0, *)
     @available(watchOS, unavailable)
     public var placeholderText: Color {
         return colorStorage["placeholderText"]!
     }
     
     /// The color to use for selected text.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -139,7 +139,7 @@ public struct SystemColor {
     }
     
     /// The color to use for the background area behind text.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -148,7 +148,7 @@ public struct SystemColor {
     }
     
     /// The color to use for the background of selected text.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -157,7 +157,7 @@ public struct SystemColor {
     }
     
     /// The color to use for the keyboard focus ring around controls.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -166,7 +166,7 @@ public struct SystemColor {
     }
     
     /// The color to use for selected text in an unemphasized context.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -175,7 +175,7 @@ public struct SystemColor {
     }
     
     /// The color to use for the text background in an unemphasized context.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -186,14 +186,14 @@ public struct SystemColor {
     // MARK: - Content
     
     /// The color to use for links.
-    @available(iOS 14.0, tvOS 14.0, macOS 10.16, *)
+    @available(iOS 14.0, tvOS 14.0, macOS 11.0, *)
     @available(watchOS, unavailable)
     public var link: Color {
         return colorStorage["link"]!
     }
     
     /// The color to use for the background of selected and emphasized content.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -202,7 +202,7 @@ public struct SystemColor {
     }
     
     /// The color to use for selected and unemphasized content.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -213,7 +213,7 @@ public struct SystemColor {
     // MARK: - Menu
     
     /// The color to use for the text in menu items.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -224,7 +224,7 @@ public struct SystemColor {
     // MARK: - Table Colors
     
     /// The color to use for the optional gridlines, such as those in a table view.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -233,7 +233,7 @@ public struct SystemColor {
     }
     
     /// The color to use for text in header cells in table views and outline views.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -242,7 +242,7 @@ public struct SystemColor {
     }
     
     /// The colors to use for alternating content, typically found in table views and collection views.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -253,7 +253,7 @@ public struct SystemColor {
     // MARK: - Controls
     
     /// The user's current accent color preference.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -262,7 +262,7 @@ public struct SystemColor {
     }
     
     /// The color to use for the flat surfaces of a control.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -271,7 +271,7 @@ public struct SystemColor {
     }
     
     /// The color to use for the background of large controls, such as scroll views or table views.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -280,7 +280,7 @@ public struct SystemColor {
     }
     
     /// The color to use for text on enabled controls.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -289,7 +289,7 @@ public struct SystemColor {
     }
     
     /// The color to use for text on disabled controls.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -298,7 +298,7 @@ public struct SystemColor {
     }
     
     /// The color to use for the face of a selected control—that is, a control that has been clicked or is being dragged.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -307,7 +307,7 @@ public struct SystemColor {
     }
     
     /// The color to use for text in a selected control—that is, a control being clicked or dragged.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -316,7 +316,7 @@ public struct SystemColor {
     }
     
     /// The color to use for text in a selected control.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -325,7 +325,7 @@ public struct SystemColor {
     }
     
     /// The patterned color to use for the background of a scrubber control.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -336,7 +336,7 @@ public struct SystemColor {
     // MARK: - Window Colors
     
     /// The color to use for the window background.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -345,7 +345,7 @@ public struct SystemColor {
     }
     
     /// The color to use for text in a window's frame.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -354,7 +354,7 @@ public struct SystemColor {
     }
     
     /// The color to use in the area beneath your window's views.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -366,7 +366,7 @@ public struct SystemColor {
     // MARK: - Highlights and Shadows
     
     /// The highlight color to use for the bubble that shows inline search result values.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -375,7 +375,7 @@ public struct SystemColor {
     }
     
     /// The color to use as a virtual light source on the screen.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -384,7 +384,7 @@ public struct SystemColor {
     }
     
     /// The color to use for virtual shadows cast by raised objects on the screen.
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)
@@ -449,7 +449,7 @@ public struct SystemColor {
     // MARK: - Separators
     
     /// The color for thin borders or divider lines that allows some underlying content to be visible.
-    @available(iOS 14.0, watchOS 6.0, tvOS 14.0, macOS 10.16, *)
+    @available(iOS 14.0, watchOS 6.0, tvOS 14.0, macOS 11.0, *)
     public var separator: Color {
         return colorStorage["separator"]!
     }
@@ -460,6 +460,14 @@ public struct SystemColor {
     @available(macOS, unavailable)
     public var opaqueSeparator: Color {
         return colorStorage["opaqueSeparator"]!
+    }
+    
+    // MARK: - Clear
+    
+    /// A clear color.
+    @available(iOS 14.0, watchOS 6.0, tvOS 14.0, macOS 11.0, *)
+    public var clear: Color {
+        return Color.clear
     }
     
     // MARK: - Initialization
@@ -600,7 +608,7 @@ public struct SystemColor {
         #endif
     }
     
-    @available(macOS 10.16, *)
+    @available(macOS 11.0, *)
     @available(iOS, unavailable)
     @available(watchOS, unavailable)
     @available(tvOS, unavailable)