+++ Sources/SolarDesignSystem/Utilities/Biometric Authentication/BiometricAuthentication+Messages.swift
@@ -26,7 +26,7 @@
 
 #if canImport(LocalAuthentication) && !os(tvOS)
 
-@available(macOS 10.16, iOS 14.0, *)
+@available(macOS 11.0, iOS 14.0, *)
 @available(watchOS, unavailable)
 @available(tvOS, unavailable)
 extension BiometricAuthentication {