+++ Sources/SolarDesignSystem/Utilities/Biometric Authentication/BiometricAuthentication.swift
@@ -28,7 +28,7 @@
 import LocalAuthentication
 
 /// A wrapper that streamlines biometric authentication.
-@available(macOS 10.16, iOS 14.0, *)
+@available(macOS 11.0, iOS 14.0, *)
 @available(watchOS, unavailable)
 @available(tvOS, unavailable)
 public class BiometricAuthentication {